+++ layers/Schema/packages/customposts/src/ModuleProcessors/FormInputs/FilterMultipleInputs.php
@@ -1,0 +1,107 @@
+<?php
+
+declare(strict_types=1);
+
+namespace PoPSchema\CustomPosts\ModuleProcessors\FormInputs;
+
+use PoP\ComponentModel\ModuleProcessors\AbstractFormInputModuleProcessor;
+use PoP\ComponentModel\ModuleProcessors\DataloadQueryArgsFilterInputModuleProcessorInterface;
+use PoP\ComponentModel\ModuleProcessors\DataloadQueryArgsSchemaFilterInputModuleProcessorInterface;
+use PoP\ComponentModel\ModuleProcessors\DataloadQueryArgsSchemaFilterInputModuleProcessorTrait;
+use PoP\ComponentModel\ModuleProcessors\FormMultipleInputModuleProcessorTrait;
+use PoP\ComponentModel\PoP_InputUtils;
+use PoP\ComponentModel\Schema\SchemaDefinition;
+use PoP\Translation\Facades\TranslationAPIFacade;
+
+class FilterMultipleInputs extends AbstractFormInputModuleProcessor implements DataloadQueryArgsFilterInputModuleProcessorInterface, DataloadQueryArgsSchemaFilterInputModuleProcessorInterface
+{
+    use DataloadQueryArgsSchemaFilterInputModuleProcessorTrait;
+    use FormMultipleInputModuleProcessorTrait;
+
+    public const MODULE_FILTERINPUT_CUSTOMPOSTDATES = 'filterinput-custompostdates';
+
+    public function getModulesToProcess(): array
+    {
+        return array(
+            [self::class, self::MODULE_FILTERINPUT_CUSTOMPOSTDATES],
+        );
+    }
+
+    public function getFilterInput(array $module): ?array
+    {
+        $filterInputs = [
+            self::MODULE_FILTERINPUT_CUSTOMPOSTDATES => [\PoPSchema\CustomPosts\FilterInputProcessor::class, \PoPSchema\CustomPosts\FilterInputProcessor::FILTERINPUT_CUSTOMPOSTDATES],
+        ];
+        return $filterInputs[$module[1]] ?? null;
+    }
+
+    public function getInputName(array $module)
+    {
+        switch ($module[1]) {
+            case self::MODULE_FILTERINPUT_CUSTOMPOSTDATES:
+                // Allow for multiple names, for multiple inputs
+                $name = $this->getName($module);
+                $names = array();
+                foreach ($this->getInputSubnames($module) as $subname) {
+                    $names[$subname] = PoP_InputUtils::getMultipleinputsName($name, $subname) . ($this->isMultiple($module) ? '[]' : '');
+                }
+                return $names;
+        }
+
+        return parent::getInputName($module);
+    }
+
+    public function getInputOptions(array $module)
+    {
+        $options = parent::getInputOptions($module);
+
+        switch ($module[1]) {
+            case self::MODULE_FILTERINPUT_CUSTOMPOSTDATES:
+                $options['subnames'] = ['from', 'to'];
+                break;
+        }
+
+        return $options;
+    }
+
+    public function getName(array $module)
+    {
+        switch ($module[1]) {
+            case self::MODULE_FILTERINPUT_CUSTOMPOSTDATES:
+                // Add a nice name, so that the URL params when filtering make sense
+                $names = array(
+                    self::MODULE_FILTERINPUT_CUSTOMPOSTDATES => 'date',
+                );
+                return $names[$module[1]];
+        }
+
+        return parent::getName($module);
+    }
+
+    public function getSchemaFilterInputType(array $module): ?string
+    {
+        $types = [
+            self::MODULE_FILTERINPUT_CUSTOMPOSTDATES => SchemaDefinition::TYPE_DATE,
+        ];
+        return $types[$module[1]] ?? null;
+    }
+
+    public function getSchemaFilterInputDescription(array $module): ?string
+    {
+        $translationAPI = TranslationAPIFacade::getInstance();
+        switch ($module[1]) {
+            case self::MODULE_FILTERINPUT_CUSTOMPOSTDATES:
+                $name = $this->getName($module);
+                $subnames = $this->getInputOptions($module)['subnames'];
+                return sprintf(
+                    $translationAPI->__('Search for posts between the \'from\' and \'to\' dates. Provide dates through params \'%s\' and \'%s\'', 'pop-posts'),
+                    \PoP\ComponentModel\PoP_InputUtils::getMultipleinputsName($name, $subnames[0]),
+                    \PoP\ComponentModel\PoP_InputUtils::getMultipleinputsName($name, $subnames[1])
+                );
+        }
+        return null;
+    }
+}
+
+
+
