--- layers/Schema/packages/migrate-customposts/migrate/library/forminputs/load.php
@@ -1,3 +1,0 @@
-<?php
-
-require_once 'filterinputs.php';
