+++ layers/Schema/packages/migrate-customposts/migrate/library/load.php
@@ -1,5 +1,4 @@
 <?php
 
 require_once 'functions/load.php';
-require_once 'forminputs/load.php';
 require_once 'filterinputs/load.php';
